+++ test/commands/change_refs_to_git_feature_branch_test.dart
@@ -453,7 +453,7 @@ void main() {
               'name: project2\n'
               'version: 1.0.0\n',
             );
-            File(p.join(project1.path, '.gg', '.gg_localize_refs_backup.json'))
+            File(p.join(project1.path, '.gg', 'gg_localize_refs_backup.json'))
               ..createSync(recursive: true)
               ..writeAsStringSync('{"project2":"^7.0.0"}');
 
@@ -475,7 +475,7 @@ void main() {
             );
 
             final backupJson = File(
-              p.join(project1.path, '.gg', '.gg_localize_refs_backup.json'),
+              p.join(project1.path, '.gg', 'gg_localize_refs_backup.json'),
             ).readAsStringSync();
             final backupMap = jsonDecode(backupJson) as Map<String, dynamic>;
 
@@ -506,7 +506,7 @@ void main() {
               'name: project2\n'
               'version: 1.0.0\n',
             );
-            File(p.join(project1.path, '.gg', '.gg_localize_refs_backup.json'))
+            File(p.join(project1.path, '.gg', 'gg_localize_refs_backup.json'))
               ..createSync(recursive: true)
               ..writeAsStringSync('{"project2":"^8.0.0"}');
 
@@ -528,7 +528,7 @@ void main() {
             );
 
             final backupJson = File(
-              p.join(project1.path, '.gg', '.gg_localize_refs_backup.json'),
+              p.join(project1.path, '.gg', 'gg_localize_refs_backup.json'),
             ).readAsStringSync();
             final backupMap = jsonDecode(backupJson) as Map<String, dynamic>;
 
@@ -575,7 +575,7 @@ void main() {
             );
 
             final backupJson = File(
-              p.join(project1.path, '.gg', '.gg_localize_refs_backup.json'),
+              p.join(project1.path, '.gg', 'gg_localize_refs_backup.json'),
             ).readAsStringSync();
             final backupMap = jsonDecode(backupJson) as Map<String, dynamic>;
 
@@ -625,7 +625,7 @@ void main() {
           );
 
           final backupJson = File(
-            p.join(project1.path, '.gg', '.gg_localize_refs_backup.json'),
+            p.join(project1.path, '.gg', 'gg_localize_refs_backup.json'),
           ).readAsStringSync();
           final backupMap = jsonDecode(backupJson) as Map<String, dynamic>;